+++ actiontext/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Clear embeds when emptying an `ActionText::RichText`.
+
+    *Joé Dupuis*
+
 *   The Trix dependency is now satisfied by a gem, `action_text-trix`, rather than vendored
     files. This allows applications to bump Trix versions independently of Rails
     releases. Effectively this also upgrades Trix to `>= 2.1.15`.